HAI-1277 Implement hanke form basic information page (#230)
Implemented basic information page of hanke form to match the new design. This included removing the start and end date controls and changing "Hankkeen vaihe" and "Työmaan koko" controls from select components to radio button groups.
