+++ src/common/components/textInput/TextInput.tsx
@@ -25,14 +25,14 @@ const TextInput: React.FC<PropTypes> = ({ name, label, disabled, tooltip, requir
       name={name}
       control={control}
       defaultValue=""
-      render={({ field: { onChange, onBlur, value } }) => (
+      render={({ field: { onChange, onBlur, value }, fieldState: { error } }) => (
         <HdsTextInput
           id={name}
           name={name}
           label={label || t(`hankeForm:labels:${name}`)}
           value={value || ''}
-          helperText={getInputErrorText(t, errors, name)}
-          invalid={!!errors[name]}
+          errorText={getInputErrorText(t, errors, name)}
+          invalid={Boolean(error)}
           onBlur={onBlur}
           onChange={onChange}
           disabled={disabled}
