+++ src/domain/hanke/edit/HankeForm.test.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import userEvent from '@testing-library/user-event';
 import { FORMFIELD, HankeDataFormState } from './types';
 import HankeForm from './HankeForm';
 import HankeFormContainer from './HankeFormContainer';
@@ -10,8 +11,6 @@ afterEach(cleanup);
 jest.setTimeout(10000);
 
 const nimi = 'test kuoppa';
-const alkuPvm = '24.03.2025';
-const loppuPvm = '25.03.2032';
 const hankkeenKuvaus = 'Tässä on kuvaus';
 const hankkeenOsoite = 'Sankaritie 3';
 /* Highly recommend to revise these tests to use typed constants like so
@@ -52,6 +51,7 @@ describe('HankeForm', () => {
   test('suunnitteluVaihde should be required when vaihe is suunnittelu', async () => {
     const handleIsDirtyChange = jest.fn();
     const handleFormClose = jest.fn();
+    const user = userEvent.setup();
 
     render(
       <HankeForm
@@ -69,26 +69,19 @@ describe('HankeForm', () => {
     fireEvent.change(screen.getByTestId(FORMFIELD.KATUOSOITE), {
       target: { value: hankkeenOsoite },
     });
-    fireEvent.change(screen.getByLabelText('Hankkeen alkupäivä', { exact: false }), {
-      target: { value: alkuPvm },
-    });
-    fireEvent.change(screen.getByLabelText('Hankkeen loppupäivä', { exact: false }), {
-      target: { value: loppuPvm },
-    });
 
-    screen.queryByText('Hankkeen Vaihe')?.click();
-    screen.queryAllByText('Suunnittelu')[0].click();
+    expect(screen.queryByText('Hankkeen suunnitteluvaihe')).not.toBeInTheDocument();
 
-    screen.getByTestId(FORMFIELD.YKT_HANKE).click();
+    await user.click(screen.getByRole('radio', { name: 'Suunnittelu' }));
+
+    await user.click(screen.getByRole('checkbox', { name: 'Hanke on YKT-hanke' }));
 
     expect(screen.getByRole('button', { name: 'Tallenna ja keskeytä' })).toBeDisabled();
 
-    screen.queryAllByText('Hankkeen suunnitteluvaihe')[0].click();
-    screen.queryAllByText('Yleis- tai hankesuunnittelu')[0].click();
+    await user.click(screen.getByText('Hankkeen suunnitteluvaihe'));
+    await user.click(screen.getByText('Yleis- tai hankesuunnittelu'));
 
-    await waitFor(() =>
-      expect(screen.getByRole('button', { name: 'Tallenna ja keskeytä' })).not.toBeDisabled()
-    );
+    expect(screen.getByRole('button', { name: 'Tallenna ja keskeytä' })).not.toBeDisabled();
   });
 
   test('Form should be populated correctly ', () => {
@@ -98,8 +91,6 @@ describe('HankeForm', () => {
           ...formData,
           [FORMFIELD.NIMI]: 'Formin nimi',
           [FORMFIELD.KUVAUS]: 'Formin kuvaus',
-          [FORMFIELD.ALKU_PVM]: '2022-11-06T00:00:00Z',
-          [FORMFIELD.LOPPU_PVM]: '2023-01-18T00:00:00Z',
         }}
         onIsDirtyChange={() => ({})}
         onFormClose={() => ({})}
@@ -111,20 +102,12 @@ describe('HankeForm', () => {
     expect(screen.getByTestId(FORMFIELD.NIMI)).toHaveValue('Formin nimi');
     expect(screen.getByTestId(FORMFIELD.KUVAUS)).toHaveValue('Formin kuvaus');
     expect(screen.getByText('Ohjelmointi')).toBeInTheDocument();
-    expect(screen.getByLabelText('Hankkeen alkupäivä', { exact: false })).toHaveValue('6.11.2022');
-    expect(screen.getByLabelText('Hankkeen loppupäivä', { exact: false })).toHaveValue('18.1.2023');
   });
 
   test('HankeFormContainer integration should work ', async () => {
     render(<HankeFormContainer />);
     fireEvent.change(screen.getByTestId(FORMFIELD.NIMI), { target: { value: nimi } });
     fireEvent.change(screen.getByTestId(FORMFIELD.KUVAUS), { target: { value: hankkeenKuvaus } });
-    fireEvent.change(screen.getByLabelText('Hankkeen alkupäivä', { exact: false }), {
-      target: { value: alkuPvm },
-    });
-    fireEvent.change(screen.getByLabelText('Hankkeen loppupäivä', { exact: false }), {
-      target: { value: loppuPvm },
-    });
     screen.queryAllByText('Hankkeen Vaihe')[0].click();
     screen.queryAllByText('Ohjelmointi')[0].click();
 
@@ -134,50 +117,5 @@ describe('HankeForm', () => {
 
     expect(screen.getByTestId(FORMFIELD.NIMI)).toHaveValue(nimi);
     expect(screen.getByTestId(FORMFIELD.KUVAUS)).toHaveValue(hankkeenKuvaus);
-  });
-
-  test('Date control validations should work', async () => {
-    render(<HankeFormContainer />);
-
-    const startDateControl = screen.getByLabelText('Hankkeen alkupäivä', { exact: false });
-    const endDateControl = screen.getByLabelText('Hankkeen loppupäivä', { exact: false });
-
-    fireEvent.change(startDateControl, {
-      target: { value: '1.13.2023' },
-    });
-    fireEvent.blur(startDateControl);
-
-    await waitFor(() =>
-      expect(screen.queryByText('Kentän tyyppi on virheellinen')).toBeInTheDocument()
-    );
-
-    fireEvent.change(startDateControl, {
-      target: { value: '1.12.2023' },
-    });
-    fireEvent.blur(startDateControl);
-
-    await waitFor(() =>
-      expect(screen.queryByText('Kentän tyyppi on virheellinen')).not.toBeInTheDocument()
-    );
-
-    fireEvent.change(endDateControl, {
-      target: { value: '1.11.2023' },
-    });
-    fireEvent.blur(endDateControl);
-
-    await waitFor(() =>
-      expect(screen.queryByText('Ensimmäinen mahdollinen päivä on 1.12.2023')).toBeInTheDocument()
-    );
-
-    fireEvent.change(endDateControl, {
-      target: { value: '1.12.2023' },
-    });
-    fireEvent.blur(endDateControl);
-
-    await waitFor(() =>
-      expect(
-        screen.queryByText('Ensimmäinen mahdollinen päivä on 1.12.2023')
-      ).not.toBeInTheDocument()
-    );
   });
 });
