+++ src/domain/hanke/edit/hankeSchema.ts
@@ -1,5 +1,3 @@
-import startOfDay from 'date-fns/startOfDay';
-import endOfDay from 'date-fns/endOfDay';
 import { $enum } from 'ts-enum-util';
 import yup from '../../../common/utils/yup';
 import { HANKE_VAIHE, HANKE_SUUNNITTELUVAIHE } from '../../types/hanke';
@@ -31,35 +29,14 @@ export const hankeSchema = yup.object().shape({
   [FORMFIELD.NIMI]: yup.string().min(3).required(),
   [FORMFIELD.KUVAUS]: yup.string().required().min(1),
   [FORMFIELD.KATUOSOITE]: yup.string().required(),
-  [FORMFIELD.ALKU_PVM]: yup.date().nullable().required(),
-  [FORMFIELD.LOPPU_PVM]: yup
-    .date()
-    .nullable()
-    .required()
-    .when(
-      FORMFIELD.ALKU_PVM,
-      // eslint-disable-next-line
-      // @ts-ignore nullable doesnt work with TS
-      (alkuPvm: Date, schema: yup.DateSchema) => {
-        try {
-          return alkuPvm ? schema.min(alkuPvm) : schema;
-        } catch (error) {
-          return schema;
-        }
-      }
-    ),
   [FORMFIELD.VAIHE]: yup.mixed().oneOf($enum(HANKE_VAIHE).getValues()).required(),
   [FORMFIELD.SUUNNITTELUVAIHE]: yup
     .mixed()
     .nullable()
     .when([FORMFIELD.VAIHE], {
       is: HANKE_VAIHE.SUUNNITTELU,
-      then: yup.mixed().oneOf($enum(HANKE_SUUNNITTELUVAIHE).getValues()),
+      then: yup.mixed().oneOf($enum(HANKE_SUUNNITTELUVAIHE).getValues()).required(),
     }),
-  [FORMFIELD.SUUNNITTELUVAIHE]: yup.string().nullable().when([FORMFIELD.VAIHE], {
-    is: 'SUUNNITTELU',
-    then: yup.string().required(),
-  }),
   [FORMFIELD.OMISTAJAT]: yup
     .array()
     .nullable()
@@ -69,32 +46,6 @@ export const hankeSchema = yup.object().shape({
     ),
   [FORMFIELD.ARVIOIJAT]: yup.array().nullable().ensure().of(contactSchema),
   [FORMFIELD.TOTEUTTAJAT]: yup.array().nullable().ensure().of(contactSchema),
-  [FORMFIELD.HAITTA_ALKU_PVM]: yup
-    .date()
-    .nullable()
-    .when(
-      ['$currentFormPage', FORMFIELD.ALKU_PVM, FORMFIELD.LOPPU_PVM],
-      // eslint-disable-next-line
-      // @ts-ignore nullable doesnt work with TS
-      (currentFormPage: number, alkuPvm: Date, loppuPvm: Date, schema: yup.DateSchema) => {
-        if (currentFormPage !== 4) return schema;
-        return alkuPvm
-          ? schema.min(startOfDay(new Date(alkuPvm))).max(endOfDay(new Date(loppuPvm)))
-          : schema;
-      }
-    ),
-  [FORMFIELD.HAITTA_LOPPU_PVM]: yup
-    .date()
-    .nullable()
-    .when(
-      ['$currentFormPage', FORMFIELD.ALKU_PVM, FORMFIELD.LOPPU_PVM],
-      // eslint-disable-next-line
-      // @ts-ignore nullable doesnt work with TS
-      (currentFormPage: number, alkuPvm: Date, loppuPvm: Date, schema: yup.DateSchema) => {
-        if (currentFormPage !== 4) return schema;
-        return loppuPvm
-          ? schema.min(startOfDay(new Date(alkuPvm))).max(endOfDay(new Date(loppuPvm)))
-          : schema;
-      }
-    ),
+  [FORMFIELD.HAITTA_ALKU_PVM]: yup.date().nullable(),
+  [FORMFIELD.HAITTA_LOPPU_PVM]: yup.date().nullable(),
 });
