+++ src/locales/fi.json
@@ -54,7 +54,8 @@
       "stringMax": "Kentän pituus oltava enintään {{max}} merkkiä",
       "dateMin": "Ensimmäinen mahdollinen päivä on {{min}}",
       "dateMax": "Viimeinen mahdollinen päivä on {{max}}"
-    }
+    },
+    "requiredInstruction": "Kaikki tähdellä * merkityt kentät ovat pakollisia."
   },
   "routes": {
     "HOME": {
@@ -310,7 +311,8 @@
     },
     "perustiedotForm": {
       "header": "Perustiedot",
-      "ytkHankeHeader": "YKT-hanke"
+      "ytkHankeHeader": "YKT-hanke",
+      "instructions": "<1>Hankkeen luonnin kautta pääset lähettämään myös hakemuksia.</1><1><0>HUOM!</0> Mikäli teet pelkkää johtoselvitystä yksityiselle alueelle, <2>täytä hakemus</2>. Yleisten alueiden johtoselvitykset haetaan hankkeen luonnin jälkeen kaivuilmoituksen kautta.</1><1>Lisäämällä hankkeen tiedot, pystyt kopioimaan ne lopuksi tarvitsemiisi hakemuksiin.</1>"
     },
     "hankkeenAlueForm": {
       "header": "Alueet",
